Restructured packages so battler classes in their own package as well.  Changed Character class to Player, since character is already a java function, and it will cause errors when it is undefined which is being referred to
