+++ src/main/java/Attacks/Bash.java
@@ -1,6 +1,5 @@
-package main.java.Attacks;
-
-import main.java.Battler;
+package Attacks;
+import Battlers.*;
 
 public class Bash extends Attack {
