+++ src/main/java/Attacks/Heal.java
@@ -1,6 +1,5 @@
-package main.java.Attacks;
-
-import main.java.Battler;
+package Attacks;
+import Battlers.*;
 
 public class Heal extends Attack {
