+++ src/main/java/Battlers/Player.java
@@ -1,29 +1,28 @@
-package main.java;
+package Battlers;
+import Attacks.*;
 
-import main.java.Attacks.*;
+public class Player extends Battler{
 
-public class Character extends Battler{
-
-    public Character(){
+    public Player(){
         super();
         addSpecialAttack(new DualSlash());
         addSpecialAttack(new Pierce());
     }
 
-    public Character(String name, int HP, int MaxHP, int MP, int MaxMP, int Atk, int Def,
-                     int CritRate, int HitRate, int EvaRate){
+    public Player(String name, int HP, int MaxHP, int MP, int MaxMP, int Atk, int Def,
+                  int CritRate, int HitRate, int EvaRate){
         super(name, HP, MaxHP, MP, MaxMP, Atk, Def, CritRate, HitRate, EvaRate);
         addSpecialAttack(new DualSlash());
         addSpecialAttack(new Pierce());
     }
 
-    public Character(String name, int HP, int MaxHP, int MP, int MaxMP, int Atk, int Def){
+    public Player(String name, int HP, int MaxHP, int MP, int MaxMP, int Atk, int Def){
         super(name, HP, MaxHP, MP, MaxMP, Atk, Def);
         addSpecialAttack(new DualSlash());
         addSpecialAttack(new Pierce());
     }
 
-    public Character(String name, int MaxHP, int MaxMP, int Atk, int Def){
+    public Player(String name, int MaxHP, int MaxMP, int Atk, int Def){
         super(name, MaxHP, MaxMP, Atk, Def);
         addSpecialAttack(new DualSlash());
         addSpecialAttack(new Pierce());
